pyproject.toml
[project]
name = "qa-guru-belevtseva-hw-21"
version = "0.1.0"
description = ""
authors = [
    {name = "Darya Belevtseva",email = "dascha067@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "allure-pytest (>=2.15.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
